chore: consolidate annotation methods
